Move functions to common crate.
