+++ src/common/storage/src/parquet_rs.rs
@@ -17,6 +17,7 @@ use common_base::runtime::execute_futures_in_parallel;
 use common_base::runtime::GLOBAL_MEM_STAT;
 use common_exception::ErrorCode;
 use common_exception::Result;
+use common_expression::FieldIndex;
 use opendal::BlockingOperator;
 use opendal::Operator;
 use parquet::arrow::parquet_to_arrow_schema;
@@ -25,13 +26,6 @@ use parquet::file::footer::decode_metadata;
 use parquet::file::metadata::ParquetMetaData;
 
 pub const FOOTER_SIZE: u64 = 8;
-
-fn convert_error(path: &str, e: parquet::errors::ParquetError) -> ErrorCode {
-    ErrorCode::BadBytes(format!(
-        "Failed to decode parquet metadata from file {}, error: {}",
-        path, e
-    ))
-}
 
 #[async_backtrace::framed]
 pub async fn read_parquet_schema_async_rs(operator: &Operator, path: &str) -> Result<ArrowSchema> {
@@ -133,13 +127,12 @@ pub async fn read_metadata_async(
     let footer = operator
         .range_read(path, (file_size - FOOTER_SIZE)..file_size)
         .await?;
-    let metadata_len = decode_footer(&footer[0..8].try_into().unwrap())
-        .map_err(|e| convert_error(path, e))? as u64;
+    let metadata_len = decode_footer(&footer[0..8].try_into().unwrap())? as u64;
     check_meta_size(file_size, metadata_len)?;
     let metadata = operator
         .range_read(path, (file_size - FOOTER_SIZE - metadata_len)..file_size)
         .await?;
-    decode_metadata(&metadata).map_err(|e| convert_error(path, e))
+    Ok(decode_metadata(&metadata)?)
 }
 
 #[allow(dead_code)]
@@ -150,12 +143,11 @@ pub fn read_metadata(
 ) -> Result<ParquetMetaData> {
     check_footer_size(file_size)?;
     let footer = operator.range_read(path, (file_size - FOOTER_SIZE)..file_size)?;
-    let metadata_len = decode_footer(&footer[0..8].try_into().unwrap())
-        .map_err(|e| convert_error(path, e))? as u64;
+    let metadata_len = decode_footer(&footer[0..8].try_into().unwrap())? as u64;
     check_meta_size(file_size, metadata_len)?;
     let metadata =
         operator.range_read(path, (file_size - FOOTER_SIZE - metadata_len)..file_size)?;
-    decode_metadata(&metadata).map_err(|e| convert_error(path, e))
+    Ok(decode_metadata(&metadata)?)
 }
 
 /// check file is large enough to hold footer
@@ -178,5 +170,125 @@ fn check_meta_size(file_size: u64, metadata_len: u64) -> Result<()> {
         )))
     } else {
         Ok(())
+    }
+}
+
+#[derive(Debug, PartialEq, Eq)]
+pub enum ParquetSchemaTreeNode {
+    Leaf(usize),
+    Inner(Vec<ParquetSchemaTreeNode>),
+}
+
+/// Convert [`parquet::schema::types::Type`] to a tree structure.
+pub fn build_parquet_schema_tree(
+    ty: &parquet::schema::types::Type,
+    leave_id: &mut usize,
+) -> ParquetSchemaTreeNode {
+    match ty {
+        parquet::schema::types::Type::PrimitiveType { .. } => {
+            let res = ParquetSchemaTreeNode::Leaf(*leave_id);
+            *leave_id += 1;
+            res
+        }
+        parquet::schema::types::Type::GroupType { fields, .. } => {
+            let mut children = Vec::with_capacity(fields.len());
+            for field in fields.iter() {
+                children.push(build_parquet_schema_tree(field, leave_id));
+            }
+            ParquetSchemaTreeNode::Inner(children)
+        }
+    }
+}
+
+/// Traverse the schema tree by `path` to collect the leaves' ids.
+pub fn traverse_parquet_schema_tree(
+    node: &ParquetSchemaTreeNode,
+    path: &[FieldIndex],
+    leaves: &mut Vec<usize>,
+) {
+    match node {
+        ParquetSchemaTreeNode::Leaf(id) => {
+            leaves.push(*id);
+        }
+        ParquetSchemaTreeNode::Inner(children) => {
+            if path.is_empty() {
+                // All children should be included.
+                for child in children.iter() {
+                    traverse_parquet_schema_tree(child, path, leaves);
+                }
+            } else {
+                let child = path[0];
+                traverse_parquet_schema_tree(&children[child], &path[1..], leaves);
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use common_expression::types::NumberDataType;
+    use common_expression::TableDataType;
+    use common_expression::TableField;
+    use common_expression::TableSchema;
+    use parquet_rs::arrow::arrow_to_parquet_schema;
+
+    use super::build_parquet_schema_tree;
+    use crate::plan::projection::ParquetSchemaTreeNode;
+
+    #[test]
+    fn test_build_parquet_schema_tree() {
+        // Test schema (6 physical columns):
+        // a: Int32,            (leave id: 0, path: [0])
+        // b: Tuple (
+        //    c: Int32,         (leave id: 1, path: [1, 0])
+        //    d: Tuple (
+        //        e: Int32,     (leave id: 2, path: [1, 1, 0])
+        //        f: String,    (leave id: 3, path: [1, 1, 1])
+        //    ),
+        //    g: String,        (leave id: 4, path: [1, 2])
+        // )
+        // h: String,           (leave id: 5, path: [2])
+        let schema = TableSchema::new(vec![
+            TableField::new("a", TableDataType::Number(NumberDataType::Int32)),
+            TableField::new("b", TableDataType::Tuple {
+                fields_name: vec!["c".to_string(), "d".to_string(), "g".to_string()],
+                fields_type: vec![
+                    TableDataType::Number(NumberDataType::Int32),
+                    TableDataType::Tuple {
+                        fields_name: vec!["e".to_string(), "f".to_string()],
+                        fields_type: vec![
+                            TableDataType::Number(NumberDataType::Int32),
+                            TableDataType::String,
+                        ],
+                    },
+                    TableDataType::String,
+                ],
+            }),
+            TableField::new("h", TableDataType::String),
+        ]);
+        let arrow_fields = schema.to_arrow().fields;
+        let arrow_schema = arrow_schema::Schema::new(
+            arrow_fields
+                .into_iter()
+                .map(arrow_schema::Field::from)
+                .collect::<Vec<_>>(),
+        );
+        let schema_desc = arrow_to_parquet_schema(&arrow_schema).unwrap();
+        let mut leave_id = 0;
+        let tree = build_parquet_schema_tree(schema_desc.root_schema(), &mut leave_id);
+        assert_eq!(leave_id, 6);
+        let expected_tree = ParquetSchemaTreeNode::Inner(vec![
+            ParquetSchemaTreeNode::Leaf(0),
+            ParquetSchemaTreeNode::Inner(vec![
+                ParquetSchemaTreeNode::Leaf(1),
+                ParquetSchemaTreeNode::Inner(vec![
+                    ParquetSchemaTreeNode::Leaf(2),
+                    ParquetSchemaTreeNode::Leaf(3),
+                ]),
+                ParquetSchemaTreeNode::Leaf(4),
+            ]),
+            ParquetSchemaTreeNode::Leaf(5),
+        ]);
+        assert_eq!(tree, expected_tree);
     }
 }
